+++ src/qutip_qoc/__init__.py
@@ -1,7 +1,7 @@
 from .version import version as __version__  # noqa
 
-from qutip_qoc.time_interval import TimeInterval
+from qutip_qoc.time import TimeInterval
 from qutip_qoc.objective import Objective
-from qutip_qoc.optimize import optimize_pulses
+from qutip_qoc.pulse_optim import optimize_pulses
 
 __all__ = ["TimeInterval", "Objective", "optimize_pulses"]
