+++ tests/test_result.py
@@ -9,9 +9,9 @@ import jax
 import jax.numpy as jnp
 import collections
 
-from qutip_qoc.optimize import optimize_pulses
+from qutip_qoc.pulse_optim import optimize_pulses
 from qutip_qoc.objective import Objective
-from qutip_qoc.time_interval import TimeInterval
+from qutip_qoc.time import TimeInterval
 from qutip_qoc.result import Result
 
 Case = collections.namedtuple(
